+++ src/app/details/details.component.ts
@@ -1,15 +1,25 @@
-import { Component } from '@angular/core';
+import {Component, inject} from '@angular/core';
+import {CommonModule} from '@angular/common';
+import {ActivatedRoute} from '@angular/router';
+import {HousingService} from '../housing.service';
+import {HousingLocation} from '../housinglocation';
 
 @Component({
   selector: 'app-details',
   imports: [],
   template: `
     <p>
-      details works!
+      details works! {{ housingLocationId }}
     </p>
   `,
   styles: ``
 })
-export class DetailsComponent {
 
+export class DetailsComponent {
+  route: ActivatedRoute = inject(ActivatedRoute);
+  housingLocationId = -1;
+
+  constructor() {
+      this.housingLocationId = Number(this.route.snapshot.params['id']);
+  }
 }